requirements.txt
fastapi
uvicorn
pandas
numpy
requests
python-multipart
dotenv
google.generativeai